Release v0.6.2: Fixed "Fast eq" path.
It doesn't seem worth it to be honest. A couple of times quicker equilibration, a couple of times slower program.

+++ gpupol2/gpupol_def.h
@@ -66,7 +66,7 @@ typedef struct SimProperties{
 	
 	uint seed;
 	int polLength;
-	int fastEq;
+	long fastEq;
 	int equilibrated;
 	char* dirOut;
 	int nDevices;
